+++ src/main/java/library/library/modelo/Libro.java
@@ -2,6 +2,9 @@ package library.library.modelo;
 
 import jakarta.persistence.*;
 
+/**
+ * Clase plantilla para los libros
+ * */
 //Registra la entidad
 @Entity
 //Table permite mapear nuestra entidad contra una tabla que tenga el mismo nombre
@@ -10,24 +13,24 @@ public class Libro {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
     private String titulo;
     private boolean alquilado;
 
     public Libro() {
     }
 
-    public Libro(long id, String titulo, boolean alquilado) {
+    public Libro(int id, String titulo, boolean alquilado) {
         this.id = id;
         this.titulo = titulo;
         this.alquilado = alquilado;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
