+++ src/main/java/library/library/persistence/entity/RoleEntity.java
@@ -11,7 +11,7 @@ public class RoleEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Enumerated(EnumType.STRING)
     private RoleEnum name;
@@ -19,16 +19,16 @@ public class RoleEntity {
     public RoleEntity() {
     }
 
-    public RoleEntity(Long id, RoleEnum name) {
+    public RoleEntity(int id, RoleEnum name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
